+++ packages/core-transactions/src/verification/index.ts
@@ -1,2 +1,2 @@
-export * from "./multi-signature-verification";
-export * from "./second-signature-verification";
+export * from "./multi-signature-verification-memoized";
+export * from "./second-signature-verification-memoized";
